t/tokenizer-test-chars.dat

#data escaped
\uD800\uD801\uDFF0\uDFFF
#errors
no DOCTYPE
#document escaped
| <html>
|   <head>
|   <body>
|     "\uD800\uD801\uDFF0\uDFFF"

#data escaped
<p>aaa\uD800\uD801\uDFF0\uDFFF
#errors
no DOCTYPE
#document escaped
| <html>
|   <head>
|   <body>
|     <p>
|       "aaa\uD800\uD801\uDFF0\uDFFF"

#data
<p>...&NotNestedGreaterGreater;?
#errors
no DOCTYPE
#document escaped
| <html>
|   <head>
|   <body>
|     <p>
|       "...\u2aa2\u0338?"

#data escaped
<p>\u0003
#errors
no DOCTYPE
U+0003
#document escaped
| <html>
|   <head>
|   <body>
|     <p>
|       "\u0003"

#data escaped
<p>\u000B
#errors
no DOCTYPE
U+000B
#document escaped
| <html>
|   <head>
|   <body>
|     <p>
|       "\u000b"
